pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "news_crawlers"
version = "1.0.0"
authors = [
    {name = "Jost Prevc", email = "jost.prevc@gmail.com"},
]
description = "An extensible python library to create web crawlers which alert users on news."
readme = "README.md"
requires-python = ">=3.7"
keywords = ["crawler", "news"]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
        "beautifulsoup4",
        "pydantic",
        "PyYAML",
        "requests",
        "numpy",
        "parameterized",
        "pytest",
    ]

[tool.setuptools]
packages = ["news_crawlers"]

[project.optional-dependencies]
dev = ["pytest", "parameterized", "pylint", "pre-commit", "black[d]", "mypy", "tox"]
test = ["pytest", "parameterized"]

[project.scripts]
news_crawlers = "news_crawlers.main:main"

[tool.black]
line-length = 120

[mypy]
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
